pyproject.toml
[tool.poetry]
name = "assign-ai"
version = "0.1.0"
description = ""
authors = ["su-shubham <code.shubham.com@gmail.com>"]
readme = "README.md"
packages = [{include = "assign_ai"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
sqlmodel = "^0.0.8"
ruff = "^0.0.261"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.6"
arel = {git = "https://github.com/florimondmanca/arel", rev = "fm/middleware"}
websockets = "^11.0.2"
python-multipart = "^0.0.6"
jwt = "^1.3.1"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
sqlalchemy-utils = "^0.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
